build with CMake
- Introduces find_package support for the various foobar2000 libs
  used in here
- Removes the included binaries and gitignores out-of-tree builds
  under folders in the project root, starting with "build"
  (for example "build/", "build64/", "buildarm64/")
