--- foo_cuefixer.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.28307.271
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "foo_cuefixer", "foo_cuefixer\foo_cuefixer.vcxproj", "{F797D00A-0EF9-4B1C-BC02-F58E60D80355}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{F797D00A-0EF9-4B1C-BC02-F58E60D80355}.Debug|x64.ActiveCfg = Debug|x64
-		{F797D00A-0EF9-4B1C-BC02-F58E60D80355}.Debug|x64.Build.0 = Debug|x64
-		{F797D00A-0EF9-4B1C-BC02-F58E60D80355}.Debug|x86.ActiveCfg = Debug|Win32
-		{F797D00A-0EF9-4B1C-BC02-F58E60D80355}.Debug|x86.Build.0 = Debug|Win32
-		{F797D00A-0EF9-4B1C-BC02-F58E60D80355}.Release|x64.ActiveCfg = Release|x64
-		{F797D00A-0EF9-4B1C-BC02-F58E60D80355}.Release|x64.Build.0 = Release|x64
-		{F797D00A-0EF9-4B1C-BC02-F58E60D80355}.Release|x86.ActiveCfg = Release|Win32
-		{F797D00A-0EF9-4B1C-BC02-F58E60D80355}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {319DC3EC-3D52-4F6C-8B6B-E15F8BDF802B}
-	EndGlobalSection
-EndGlobal
